move class off file
